refactor(decompose): split webapp/e2e-tests/electron/for_feature_development_not_LT_verification/terminals/lifecycle/electron-slider-click-spawn.spec.ts into folder
