+++ webapp/e2e-tests/electron/for_feature_development_not_LT_verification/terminals/lifecycle/electron-slider-click-spawn.spec.ts
@@ -28,233 +28,18 @@
  * mousedown handlers that cause the bug.
  */
 
-import { test as base, expect, _electron as electron } from '@playwright/test';
-import type { ElectronApplication, Page } from '@playwright/test';
-import * as path from 'path';
-import * as fs from 'fs/promises';
-import * as os from 'os';
-import type { Core as CytoscapeCore } from 'cytoscape';
-import type { ElectronAPI } from '@/shell/electron';
-
-const PROJECT_ROOT = path.resolve(process.cwd());
-const FIXTURE_VAULT_PATH = path.join(PROJECT_ROOT, 'example_folder_fixtures', 'example_small');
-
-interface ExtendedWindow {
-  cytoscapeInstance?: CytoscapeCore;
-  electronAPI?: ElectronAPI;
-}
-
-const test = base.extend<{
-  electronApp: ElectronApplication;
-  appWindow: Page;
-}>({
-  electronApp: async ({}, use) => {
-    const tempUserDataPath = await fs.mkdtemp(path.join(os.tmpdir(), 'voicetree-slider-click-test-'));
-
-    // Write config to auto-load test vault
-    const configPath = path.join(tempUserDataPath, 'voicetree-config.json');
-    await fs.writeFile(configPath, JSON.stringify({
-      lastDirectory: FIXTURE_VAULT_PATH,
-      suffixes: {
-        [FIXTURE_VAULT_PATH]: ''
-      }
-    }, null, 2), 'utf8');
-
-    // Write settings with known contextNodeMaxDistance
-    const settingsPath = path.join(tempUserDataPath, 'voicetree-settings.json');
-    await fs.writeFile(settingsPath, JSON.stringify({
-      contextNodeMaxDistance: 5
-    }, null, 2), 'utf8');
-    console.log('[Test] Created config with contextNodeMaxDistance: 5');
-
-    const electronApp = await electron.launch({
-      args: [
-        path.join(PROJECT_ROOT, 'dist-electron/main/index.js'),
-        `--user-data-dir=${tempUserDataPath}`
-      ],
-      env: {
-        ...process.env,
-        NODE_ENV: 'test',
-        HEADLESS_TEST: '1',
-        MINIMIZE_TEST: '1',
-        VOICETREE_PERSIST_STATE: '1'
-      },
-      timeout: 10000
-    });
-
-    await use(electronApp);
-
-    // Graceful shutdown
-    try {
-      const window = await electronApp.firstWindow();
-      await window.evaluate(async () => {
-        const api = (window as unknown as ExtendedWindow).electronAPI;
-        if (api) {
-          await api.main.stopFileWatching();
-        }
-      });
-      await window.waitForTimeout(300);
-    } catch {
-      console.log('Note: Could not stop file watching during cleanup');
-    }
-
-    await electronApp.close();
-    await fs.rm(tempUserDataPath, { recursive: true, force: true });
-  },
-
-  appWindow: async ({ electronApp }, use) => {
-    const window = await electronApp.firstWindow({ timeout: 15000 });
-
-    window.on('console', msg => {
-      console.log(`BROWSER [${msg.type()}]:`, msg.text());
-    });
-
-    window.on('pageerror', error => {
-      console.error('PAGE ERROR:', error.message);
-    });
-
-    await window.waitForLoadState('domcontentloaded');
-
-    try {
-      await window.waitForFunction(() => (window as unknown as ExtendedWindow).cytoscapeInstance, { timeout: 10000 });
-    } catch (error) {
-      console.error('Failed to initialize cytoscape instance:', error);
-      throw error;
-    }
-
-    await window.waitForTimeout(1000);
-    await use(window);
-  }
-});
-
-/**
- * Wait for graph to load with nodes
- */
-async function waitForGraphLoaded(appWindow: Page): Promise<void> {
-  await expect.poll(async () => {
-    return appWindow.evaluate(() => {
-      const cy = (window as unknown as ExtendedWindow).cytoscapeInstance;
-      if (!cy) return 0;
-      return cy.nodes().length;
-    });
-  }, {
-    message: 'Waiting for graph to load nodes',
-    timeout: 15000,
-    intervals: [500, 1000, 1000]
-  }).toBeGreaterThan(0);
-}
-
-/**
- * Find non-context node IDs for testing (returns multiple for regression test)
- */
-async function getNonContextNodeIds(appWindow: Page, count: number = 2): Promise<string[]> {
-  return appWindow.evaluate((n) => {
-    const cy = (window as unknown as ExtendedWindow).cytoscapeInstance;
-    if (!cy) throw new Error('Cytoscape not available');
-
-    // Find nodes that are NOT context nodes and have a file extension (markdown nodes)
-    const nodes = cy.nodes().filter((node) => {
-      const id = node.id();
-      const isContextNode = node.data('isContextNode') === true;
-      const hasFileExtension = /\.\w+$/.test(id);
-      return !isContextNode && hasFileExtension;
-    });
-
-    if (nodes.length < n) {
-      throw new Error(`Not enough non-context nodes found (need ${n}, found ${nodes.length})`);
-    }
-
-    return nodes.slice(0, n).map(node => node.id());
-  }, count);
-}
-
-/**
- * Get the SCREEN position of a cytoscape node (accounting for pan, zoom, and container offset).
- * This is where the node appears on the actual screen, for real mouse movements.
- */
-async function getNodeScreenPosition(appWindow: Page, nodeId: string): Promise<{x: number; y: number}> {
-  return appWindow.evaluate((id) => {
-    const cy = (window as unknown as ExtendedWindow).cytoscapeInstance;
-    if (!cy) throw new Error('Cytoscape not available');
-    const node = cy.getElementById(id);
-    // renderedPosition gives the position in the canvas coordinate system
-    const rendered = node.renderedPosition();
-    // Get the cytoscape container's position on screen
-    const container = cy.container();
-    if (!container) throw new Error('No container');
-    const rect = container.getBoundingClientRect();
-    return {
-      x: rect.left + rendered.x,
-      y: rect.top + rendered.y
-    };
-  }, nodeId);
-}
-
-/**
- * Hover over a node using REAL mouse coordinates (triggers DOM events properly)
- */
-async function hoverOverNodeReal(appWindow: Page, nodeId: string): Promise<void> {
-  const pos = await getNodeScreenPosition(appWindow, nodeId);
-  console.log(`[hoverOverNodeReal] Moving mouse to node ${nodeId} at (${pos.x}, ${pos.y})`);
-  await appWindow.mouse.move(pos.x, pos.y);
-}
-
-/**
- * Leave a node by moving mouse away
- */
-async function leaveNodeReal(appWindow: Page): Promise<void> {
-  // Move to a corner of the screen (away from any nodes)
-  await appWindow.mouse.move(10, 10);
-}
-
-/**
- * Get the floating slider element (appended to .cy-floating-overlay)
- */
-function getSlider(appWindow: Page) {
-  return appWindow.locator('.cy-floating-overlay .distance-slider').first();
-}
-
-/**
- * Get all square elements in the slider
- */
-function getSliderSquares(appWindow: Page) {
-  return appWindow.locator('.cy-floating-overlay .distance-slider > div:last-child > div');
-}
-
-/**
- * Count terminal shadow nodes in the graph
- */
-async function getTerminalCount(appWindow: Page): Promise<number> {
-  return appWindow.evaluate(() => {
-    const cy = (window as unknown as ExtendedWindow).cytoscapeInstance;
-    if (!cy) return 0;
-    return cy.nodes().filter(node =>
-      node.data('isShadowNode') === true &&
-      node.data('windowType') === 'Terminal'
-    ).length;
-  });
-}
-
-/**
- * Get the attached node ID from the most recently created terminal
- */
-async function getLastTerminalAttachedNodeId(appWindow: Page): Promise<string | null> {
-  return appWindow.evaluate(() => {
-    const cy = (window as unknown as ExtendedWindow).cytoscapeInstance;
-    if (!cy) return null;
-
-    const terminals = cy.nodes().filter(node =>
-      node.data('isShadowNode') === true &&
-      node.data('windowType') === 'Terminal'
-    );
-
-    if (terminals.length === 0) return null;
-
-    // Return the attachedToContextNodeId of the last (most recent) terminal
-    const lastTerminal = terminals.last();
-    return lastTerminal.data('attachedToNodeId') ?? null;
-  });
-}
+import { expect } from '@playwright/test';
+import {
+  getLastTerminalAttachedNodeId,
+  getNonContextNodeIds,
+  getSlider,
+  getSliderSquares,
+  getTerminalCount,
+  hoverOverNodeReal,
+  leaveNodeReal,
+  test,
+  waitForGraphLoaded
+} from './electron-slider-click-spawn/test-helpers';
 
 test.describe('Distance Slider Click Spawns Agent', () => {
